refactor(dashboard): extract shared extractCodeLocation to normalize.ts
Three files (logs, keeper-tool-call-inspector, session-trace-entry)
had identical file-path + line extraction logic (file_path/path/file +
line/line_start/lineno). Extracted to normalize.ts as extractCodeLocation
with CodeLocation interface. Removed 3 local implementations and 2
redundant type definitions (LogCodeLocation, TraceCodeLocation).
